# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Computer‑Based Testing API with subscriptions, academic management, and CBT modules.",
        "title": "CBT API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@cbt-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/class-arms": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a class arm (e.g., \"Section A\") for a school.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Academic"
                ],
                "summary": "Create a new class arm",
                "parameters": [
                    {
                        "description": "Class arm details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateClassArmRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "message + data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/class-arms/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a single class arm.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Academic"
                ],
                "summary": "Get class arm by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Class Arm ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Modify an existing class arm.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Academic"
                ],
                "summary": "Update a class arm",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Class Arm ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateClassArmRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message + data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soft‑delete a class arm.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Academic"
                ],
                "summary": "Delete a class arm",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Class Arm ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/school-class-arms/{schoolId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List all class arms belonging to a school.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Academic"
                ],
                "summary": "Get all class arms for a school",
                "parameters": [
                    {
                        "type": "string",
                        "description": "School ID",
                        "name": "schoolId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data (list)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateClassArmRequest": {
            "type": "object",
            "required": [
                "name",
                "school_id"
            ],
            "properties": {
                "arm_code": {
                    "type": "string"
                },
                "capacity": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "room_number": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateClassArmRequest": {
            "type": "object",
            "properties": {
                "arm_code": {
                    "type": "string"
                },
                "capacity": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "room_number": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter **Bearer \u0026lt;token\u0026gt;**. Example: `Bearer eyJhbGciOiJIUzI1NiIs...`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}